Remove scala.quoted.util.Var from the library
This is a nice abstration but it is not ablolutely required.
It would fit nicely in an external library.
