+++ tests/run-staging/staged-streams_1.scala
@@ -680,3 +680,22 @@ object Test {
     println(run(test10()))
   }
 }
+
+sealed trait Var[T] {
+  def get(using qctx: QuoteContext): Expr[T]
+  def update(e: Expr[T])(using qctx: QuoteContext): Expr[Unit]
+}
+
+object Var {
+  def apply[T: Type, U: Type](init: Expr[T])(body: Var[T] => Expr[U])(using qctx: QuoteContext): Expr[U] = '{
+    var x = $init
+    ${
+      body(
+        new Var[T] {
+          def get(using qctx: QuoteContext): Expr[T] = 'x
+          def update(e: Expr[T])(using qctx: QuoteContext): Expr[Unit] = '{ x = $e }
+        }
+      )
+    }
+  }
+}
